git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@13993 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/jd/gui/swing/jdgui/components/speedmeter/SpeedMeterPanel.java
@@ -81,7 +81,7 @@ public class SpeedMeterPanel extends Graph {
 
     @Override
     public int getValue() {
-        return (int) DownloadWatchDog.getInstance().getConnectionManager().getIncommingBandwidthUsage();
+        return DownloadWatchDog.getInstance().getConnectionManager().getIncommingBandwidthUsage();
     }
 
 }
